requirements.txt
streamlit 
openai-whisper 
google-genai
imageio[ffmpeg]
